# Decrease deserialization complexity from quadratic to linear