--- a/src/de.rs
+++ b/src/de.rs
@@ -216,6 +216,7 @@ impl<'de, 'b> de::Deserializer<'de> for &'b mut Deserializer<'de> {
     {
         let mut tables = self.tables()?;
         let table_indices = build_table_indices(&tables);
+        let table_pindices = build_table_pindices(&tables);
 
         let res = visitor.visit_map(MapVisitor {
             values: Vec::new().into_iter(),
@@ -225,6 +226,7 @@ impl<'de, 'b> de::Deserializer<'de> for &'b mut Deserializer<'de> {
             cur_parent: 0,
             max: tables.len(),
             table_indices: &table_indices,
+            table_pindices: &table_pindices,
             tables: &mut tables,
             array: false,
             de: self,
@@ -330,6 +332,19 @@ fn build_table_indices<'de>(tables: &[Table<'de>]) -> HashMap<Vec<Cow<'de, str>>
     res
 }
 
+fn build_table_pindices<'de>(tables: &[Table<'de>]) -> HashMap<Vec<Cow<'de, str>>, Vec<usize>> {
+    let mut res = HashMap::new();
+    for (i, table) in tables.iter().enumerate() {
+        let header = table.header.iter().map(|v| v.1.clone()).collect::<Vec<_>>();
+        for len in 0..=header.len() {
+            res.entry(header[..len].to_owned())
+                .or_insert(Vec::new())
+                .push(i);
+        }
+    }
+    res
+}
+
 fn headers_equal<'a, 'b>(hdr_a: &[(Span, Cow<'a, str>)], hdr_b: &[(Span, Cow<'b, str>)]) -> bool {
     if hdr_a.len() != hdr_b.len() {
         return false;
@@ -352,6 +367,7 @@ struct MapVisitor<'de, 'b> {
     cur_parent: usize,
     max: usize,
     table_indices: &'b HashMap<Vec<Cow<'de, str>>, Vec<usize>>,
+    table_pindices: &'b HashMap<Vec<Cow<'de, str>>, Vec<usize>>,
     tables: &'b mut [Table<'de>],
     array: bool,
     de: &'b mut Deserializer<'de>,
@@ -377,20 +393,27 @@ impl<'de, 'b> de::MapAccess<'de> for MapVisitor<'de, 'b> {
             }
 
             let next_table = {
-                let prefix = &self.tables[self.cur_parent].header[..self.depth];
-                self.tables[self.cur..self.max]
+                let prefix_stripped = self.tables[self.cur_parent].header[..self.depth]
                     .iter()
-                    .enumerate()
-                    .find(|&(_, t)| {
-                        if t.values.is_none() {
-                            return false;
+                    .map(|v| v.1.clone())
+                    .collect::<Vec<_>>();
+                self.table_pindices
+                    .get(&prefix_stripped)
+                    .and_then(|entries| {
+                        let start = entries
+                            .binary_search(&self.cur)
+                            .unwrap_or_else(std::convert::identity);
+                        if start == entries.len() || entries[start] < self.cur {
+                            return None;
                         }
-                        match t.header.get(..self.depth) {
-                            Some(header) => headers_equal(&header, &prefix),
-                            None => false,
-                        }
+                        entries[start..]
+                            .iter()
+                            .copied()
+                            .filter(|i| *i < self.max)
+                            .map(|i| (i, &self.tables[i]))
+                            .find(|(_, table)| table.values.is_some())
+                            .map(|p| p.0)
                     })
-                    .map(|(i, _)| i + self.cur)
             };
 
             let pos = match next_table {
@@ -483,6 +506,7 @@ impl<'de, 'b> de::MapAccess<'de> for MapVisitor<'de, 'b> {
             max: self.max,
             array,
             table_indices: &*self.table_indices,
+            table_pindices: &*self.table_pindices,
             tables: &mut *self.tables,
             de: &mut *self.de,
         });
@@ -546,6 +570,7 @@ impl<'de, 'b> de::SeqAccess<'de> for MapVisitor<'de, 'b> {
             cur: 0,
             array: false,
             table_indices: &*self.table_indices,
+            table_pindices: &*self.table_pindices,
             tables: &mut self.tables,
             de: &mut self.de,
         })?;