Fixes #342.

In particular, see my comment https://github.com/alexcrichton/toml-rs/issues/342#issuecomment-546190394

Values that I recorded on my machine for running the code `measure_time(n, |i| format!("[header_no_{}]\n", i))` function for varying `n`:

| benchmark  | optimizations | before this PR | after this PR |
|------------|---------------|----------------|---------------|
|          1k|             no|           170ms|           48ms|
|         10k|             no|         14 191ms|          483ms|
|        100k|             no|             n/a|         5 077ms|
|          1k|            yes|             5ms|            4ms|
|         10k|            yes|           311ms|           39ms|
|        100k|            yes|         63 850ms|          364ms|

You can nicely see how before this PR, a 10x increase in data meant a 100x increase in time spent, while after the PR it only means a 10x increase.