@alexcrichton 2019-10-28
It looks like the tests added here [may be causing a spurious failure on CI](https://github.com/alexcrichton/toml-rs/runs/277767964#step:6:132)?